refactor(dashboard): route provider hook imports through barrel
Switches four layout consumers (command-palette, refresh-control,
theme-toggle, header) to import from '@/providers' instead of the
individual provider modules. Toaster keeps its deep import because
the providers barrel imports Toaster, so routing it through the
barrel would create a cycle.
