+++ dashboard/src/components/layout/command-palette.tsx
@@ -25,9 +25,7 @@ import {
   CommandSeparator,
   CommandShortcut,
 } from "@/components/ui/command";
-import { useCommandPalette } from "@/providers/command-palette-provider";
-import { type RefreshOption, useRefreshInterval } from "@/providers/refresh-interval-provider";
-import { useTheme } from "@/providers/theme-provider";
+import { type RefreshOption, useCommandPalette, useRefreshInterval, useTheme } from "@/providers";
 
 interface NavCmd {
   label: string;
